refact blanks done
